solver: split llbop_build.go to llbop pkg
